renomeia e move funcoes e arquivos
